[Refactor] rename HiveMetastoreThriftClient to HiveMetastoreClient
