+++ fe/fe-core/src/main/java/com/starrocks/connector/iceberg/hive/HiveClientPool.java
@@ -16,10 +16,10 @@
 package com.starrocks.connector.iceberg.hive;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.starrocks.connector.hive.HiveMetaStoreThriftClient;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.metastore.HiveMetaHookLoader;
+import org.apache.hadoop.hive.metastore.HiveMetaStoreClient;
 import org.apache.hadoop.hive.metastore.IMetaStoreClient;
 import org.apache.hadoop.hive.metastore.RetryingMetaStoreClient;
 import org.apache.hadoop.hive.metastore.api.MetaException;
@@ -49,7 +49,7 @@ public class HiveClientPool extends ClientPoolImpl<IMetaStoreClient, TException>
         try {
             try {
                 return GET_CLIENT.invoke(hiveConf, (HiveMetaHookLoader) tbl -> null,
-                        HiveMetaStoreThriftClient.class.getName());
+                        HiveMetaStoreClient.class.getName());
             } catch (RuntimeException e) {
                 // any MetaException would be wrapped into RuntimeException during reflection, so let's double-check type here
                 if (e.getCause() instanceof MetaException) {
