+++ fe/fe-core/src/test/java/com/starrocks/connector/hive/HiveMetaClientTest.java
@@ -22,6 +22,7 @@ import mockit.Mocked;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.metastore.HiveMetaHookLoader;
+import org.apache.hadoop.hive.metastore.HiveMetaStoreClient;
 import org.apache.hadoop.hive.metastore.IMetaStoreClient;
 import org.apache.hadoop.hive.metastore.RetryingMetaStoreClient;
 import org.apache.hadoop.hive.metastore.api.MetaException;
@@ -41,7 +42,7 @@ import java.util.concurrent.TimeUnit;
 
 public class HiveMetaClientTest {
     @Test
-    public void testClientPool(@Mocked HiveMetaStoreThriftClient metaStoreClient) throws Exception {
+    public void testClientPool(@Mocked HiveMetaStoreClient metaStoreClient) throws Exception {
         new Expectations() {
             {
                 metaStoreClient.getTable(anyString, anyString);
@@ -104,7 +105,7 @@ public class HiveMetaClientTest {
     }
 
     @Test
-    public void testRecyclableClient(@Mocked HiveMetaStoreThriftClient metaStoreClient) throws TException {
+    public void testRecyclableClient(@Mocked HiveMetaStoreClient metaStoreClient) throws TException {
         new Expectations() {
             {
                 metaStoreClient.getTable(anyString, anyString);
